requirements.txt
opencv-python
numpy
pillow
shapely
scipy
matplotlib
tqdm